OneDrive - QuidelOrtho/SUNIL/EAG/Session 8/code1/pyproject.toml
[project]
name = "cursor-agent"
version = "0.1.0"
description = "Multi-layer Cursor Agent with Perception-Memory-Decision-Action architecture"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core
    "pydantic>=2.10.5",
    "python-dotenv>=1.0.0",
    
    # Web & APIs
    "fastapi>=0.120.4",
    "uvicorn>=0.38.0",
    "requests>=2.32.3",
    "httpx>=0.28.0",
    "aiohttp>=3.9.0",
    
    # Google APIs
    "google-api-python-client>=2.155.0",
    "google-auth-oauthlib>=1.2.1",
    "google-auth-httplib2>=0.2.0",
    
    # Telegram
    "python-telegram-bot>=22.0",
    
    # AI & ML
    "google-genai>=1.0.0",
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    
    # Content Extraction
    "trafilatura>=1.12.0",
    "pymupdf>=1.23.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    
    # Screenshot
    "playwright>=1.40.0",
    "pillow>=10.0.0",
    
    # MCP Protocol
    "mcp>=0.9.0",
    
    # Utilities
    "numpy>=1.24.0",
    "aiofiles>=23.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["agent", "mcp_servers"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
